music algorithm + coins ignore self collision

+++ Assets/Scripts/PoolManager.cs
@@ -13,12 +13,13 @@ public class PoolManager : MonoBehaviour
         public Color color;
     }
 
-
     public GameObject prefab;
     public Coin[] coins;
     public int maxCoins;
     public List<GameObject> allCoins;
     public List<GameObject> toUseCoins;
+    public float rate;
+    bool canLoad = false;
 
     // Use this for initialization
     void Awake()
@@ -35,31 +36,42 @@ public class PoolManager : MonoBehaviour
                 g.SetActive(false);
             }
         }
+        InvokeRepeating ("GetsRandomCoin", 2f, rate);
     }
 
     GameObject PopulateCoinsPrefabs(Coin c)
     {
         GameObject i = prefab;
         CoinManager cm = i.GetComponent<CoinManager>();
+        cm.ResetPosition();
         cm.speed = c.speed;
         cm.color = c.color;
         cm.value = c.value;
         return i;
     }
 
-    public GameObject GetsRandomCoin()
-    {
-        if (allCoins.Count <= 0)
-        {
-            allCoins = new List<GameObject>(toUseCoins);
-            toUseCoins.RemoveRange(0, toUseCoins.Count);
+    public void GetsRandomCoin()
+    {   
+        StartCoroutine (loadCoin (Random.Range(0, 10)));
+        if (canLoad) {
+            if (allCoins.Count <= 0)
+            {
+                allCoins = new List<GameObject>(toUseCoins);
+                toUseCoins.RemoveRange(0, toUseCoins.Count);
+            }
+
+            int value = Random.Range(0, allCoins.Count);
+            GameObject g = allCoins[value];
+            allCoins.RemoveAt(value);
+            toUseCoins.Add(g);
+            g.SetActive(true);
+            canLoad = false;
         }
+    }
 
-        int value = Random.Range(0, allCoins.Count);
-        GameObject g = allCoins[value];
-        allCoins.RemoveAt(value);
-        toUseCoins.Add(g);
-
-        return g;
+    IEnumerator loadCoin(int f)
+    {
+        yield return new WaitForSeconds (f);
+        canLoad = true;
     }
 }
